chore: move types and Visitor interface to ast
