+++ src/main/java/de/dhbw/compiler/ast/Method.java
@@ -1,7 +1,6 @@
 package de.dhbw.compiler.ast;
 
 import de.dhbw.compiler.ast.statements.Statement;
-import de.dhbw.compiler.codegeneration.Type;
 import lombok.Getter;
 
 import java.util.List;
