+++ src/main/java/de/dhbw/compiler/ast/Type.java
@@ -1,4 +1,4 @@
-package de.dhbw.compiler.codegeneration;
+package de.dhbw.compiler.ast;
 
 public interface Type {
     String getName();
