+++ src/main/java/de/dhbw/compiler/ast/Visitor.java
@@ -1,0 +1,7 @@
+package de.dhbw.compiler.ast;
+
+import de.dhbw.compiler.codegeneration.MethodCodeVisitor;
+
+public interface Visitor {
+    void accept(MethodCodeVisitor visitor);
+}
