+++ src/main/java/de/dhbw/compiler/ast/expressions/Expression.java
@@ -1,7 +1,7 @@
 package de.dhbw.compiler.ast.expressions;
 
-import de.dhbw.compiler.codegeneration.Type;
-import de.dhbw.compiler.codegeneration.Visitor;
+import de.dhbw.compiler.ast.Type;
+import de.dhbw.compiler.ast.Visitor;
 
 public sealed interface Expression extends Visitor permits
     This, Super, LocalOrFieldVar, InstVar, Unary, Binary, JInteger, JBoolean, JCharacter, JString, JNull, StmtExprExpr {
