+++ src/main/java/de/dhbw/compiler/ast/statements/If.java
@@ -2,7 +2,7 @@ package de.dhbw.compiler.ast.statements;
 
 import de.dhbw.compiler.ast.expressions.Expression;
 import de.dhbw.compiler.codegeneration.MethodCodeVisitor;
-import de.dhbw.compiler.codegeneration.Type;
+import de.dhbw.compiler.ast.Type;
 
 public non-sealed class If implements Statement {
