+++ src/main/java/de/dhbw/compiler/ast/stmtexprs/StatementExpression.java
@@ -1,7 +1,7 @@
 package de.dhbw.compiler.ast.stmtexprs;
 
 import de.dhbw.compiler.codegeneration.MethodCodeVisitor;
-import de.dhbw.compiler.codegeneration.Type;
+import de.dhbw.compiler.ast.Type;
 
 public sealed interface StatementExpression permits Assign, MethodCall, New {
