--- src/main/java/de/dhbw/compiler/codegeneration/Visitor.java
@@ -1,5 +1,0 @@
-package de.dhbw.compiler.codegeneration;
-
-public interface Visitor {
-    void accept(MethodCodeVisitor visitor);
-}
