+++ src/main/java/de/dhbw/compiler/typecheck/StatementExpressionChecker.java
@@ -9,7 +9,7 @@ import de.dhbw.compiler.ast.stmtexprs.Assign;
 import de.dhbw.compiler.ast.stmtexprs.MethodCall;
 import de.dhbw.compiler.ast.stmtexprs.New;
 import de.dhbw.compiler.ast.stmtexprs.StatementExpression;
-import de.dhbw.compiler.codegeneration.ObjectType;
+import de.dhbw.compiler.ast.ObjectType;
 import lombok.RequiredArgsConstructor;
 import lombok.Setter;
